Refactor page counter slice to include a reset action
Refactored the pageCounterSlice in the src/domain/page-counter-slice.ts file to include a new reset action. This action sets the pageCounter state to 0. This change was made to improve the functionality of the page counter.

Reset page counter when filtering by favorites

In the src/ui/components/molecules/filter-box/index.tsx file, added a condition to reset the page counter when the filter option is set to "Favorites". This ensures that the page counter starts from the first page when filtering by favorites. This change improves the user experience when filtering the results.